requirements.txt

transformers==4.41.1
requests==2.31.0
python-telegram-bot==13.15
